Factor out handler and middleware expansion utilties from the inference route
